# backend/examinationtype/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This App is used to control the examinationType. It can:\n- Add a new examinationType\n- Get the examinationType\n- Delete the examinationType\n- Modify the examinationType",
        "title": "Huovet examinationType API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5010",
    "basePath": "/",
    "paths": {
        "/examinationtype/addExaminationtype": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Insert new Examinationtype to the Examinationtype table in Huovet schema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinationtype"
                ],
                "summary": "Add new Examinationtype to database",
                "parameters": [
                    {
                        "description": "query body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExaminationtypeAddJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExaminationtypeDB"
                        }
                    },
                    "401": {
                        "description": "Unauthentication",
                        "schema": {
                            "$ref": "#/definitions/models.UnathenticatedResponse"
                        }
                    }
                }
            }
        },
        "/examinationtype/deleteExaminationtype/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove all the information of the Examinationtype.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinationtype"
                ],
                "summary": "Delete Examinationtype information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Examinationtype ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete success",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Deleting fail",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteFAilResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthentication",
                        "schema": {
                            "$ref": "#/definitions/models.UnathenticatedResponse"
                        }
                    }
                }
            }
        },
        "/examinationtype/getAllExaminationtype/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Return the all Examinationtype and releated breeds",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinationtype"
                ],
                "summary": "Returns all Examinationtype",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExaminationtypeDB"
                        }
                    },
                    "400": {
                        "description": "Seacring fail",
                        "schema": {
                            "$ref": "#/definitions/models.SearchFailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthentication",
                        "schema": {
                            "$ref": "#/definitions/models.UnathenticatedResponse"
                        }
                    }
                }
            }
        },
        "/examinationtype/getExaminationtype/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Return the information of the Examinationtype.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinationtype"
                ],
                "summary": "Return Examinationtype information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Examinationtype ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExaminationtypeDB"
                        }
                    },
                    "400": {
                        "description": "Seacring fail",
                        "schema": {
                            "$ref": "#/definitions/models.SearchFailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthentication",
                        "schema": {
                            "$ref": "#/definitions/models.UnathenticatedResponse"
                        }
                    }
                }
            }
        },
        "/examinationtype/modifyExaminationtype": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the information of the Examinationtype.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinationtype"
                ],
                "summary": "Update Examinationtype information",
                "parameters": [
                    {
                        "description": "query body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExaminationtypeDB"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update success",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthentication",
                        "schema": {
                            "$ref": "#/definitions/models.UnathenticatedResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.DeleteFAilResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "deleting fail"
                }
            }
        },
        "models.DeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "delete success"
                }
            }
        },
        "models.ExaminationtypeAddJSON": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "examination"
                },
                "location": {
                    "type": "string",
                    "example": "Tampere"
                },
                "name": {
                    "type": "string",
                    "example": "check"
                },
                "price": {
                    "type": "number",
                    "example": 9.2
                }
            }
        },
        "models.ExaminationtypeDB": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "examination"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string",
                    "example": "Tampere"
                },
                "name": {
                    "type": "string",
                    "example": "check"
                },
                "price": {
                    "type": "number",
                    "example": 9.2
                }
            }
        },
        "models.SearchFailResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "searching fail"
                }
            }
        },
        "models.UnathenticatedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "unauthenticated"
                }
            }
        },
        "models.UpdateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "update success"
                }
            }
        }
    }
}